feat: 참가 취소 버튼 분기 추가 및 컴포넌트 분리
