+++ src/components/detail/Detail.jsx
@@ -5,58 +5,29 @@ import styled from "@emotion/styled";
 import { __getDetail } from "../../redux/modules/details/detailSlice";
 import { __postApplication } from "../../redux/modules/application/applicationSlice";
 import { RiMapPin2Fill } from "react-icons/ri";
-import applicateBtnIcon from "../../assets/detailIcon/applicateBtnIcon.svg";
-import profileDefault from "../../assets/profileImg/profile_default.svg";
 
-import ButtonBasic from "../elements/ButtonBasic";
 import DetailSpan from "./elements/DetailSpan";
 import DetailMoreButton from "./elements/DetailMoreButton";
-import DetailPostingOptionModal from "./modals/DetailPostingOptionModal";
-import DetailCommentModal from "./modals/DetailCommentModal";
-import DetailApplicationList from "./DetailApplicationList";
-import DetailCommentForm from "./DetailCommentForm";
 import DetailApplicationBody from "./DetailApplicationBody";
 import { getCookies } from "../../core/cookie";
 // import { longTitleForm, titleForm } from "../../utils/editedData";
 import { Spinners } from "../../shared/layout/Spinners";
 import { __getComment } from "../../redux/modules/details/commentSlice";
 import useBuyLocation from "../../hooks/useBuyLocation";
+import DetailApplicationBtns from "./DetailApplicationBtns";
+import DetailCommentList from "./DetailCommentList";
+import DetailCreatorProfile from "./DetailCreatorProfile";
+import DetailContent from "./DetailContent";
 
 const Detail = () => {
   const dispatch = useDispatch();
   const navigate = useNavigate();
+  const tokenValue = getCookies("id");
   const postingId = parseInt(useParams().postingId);
+  const { isLoading, error } = useSelector((state) => state.getDetail);
   const details = useSelector((state) => state.getDetail.getDetail);
   const isSuccess = useSelector((state) => state.comments.isSuccess);
-  /*   const { isLoading, error } = useSelector(
-    (state) => state.getDetail.getDetail
-  ); */
-  const [postingModal, setPostingModal] = useState(false);
-  const [commentModal, setCommentModal] = useState(false);
-  const [applicationModal, setApplicationModal] = useState(false);
-  const tokenValue = getCookies("id");
-  const commentsLength = details.commentList?.length || 0;
 
-  const onClickPostingModalHandler = () => {
-    setPostingModal(!postingModal);
-  };
-  const onClickCommentModalHandler = (e) => {
-    //console.log(e.target.id);
-    if (e.target.id === details.commentList.commentId) {
-      setCommentModal(!commentModal);
-    }
-  };
-  const onClickApplicationModalHandler = () => {
-    setApplicationModal(!applicationModal);
-  };
-  const onClickApplicateHandler = () => {
-    if (!tokenValue) {
-      alert("로그인 해주세요.");
-      navigate("/login");
-    } else {
-      dispatch(__postApplication(postingId));
-    }
-  };
   const onClickMoveProfileHandler = (memberId) => {
     navigate(`/profile/${memberId}`);
   };
@@ -67,8 +38,9 @@ const Detail = () => {
     dispatch(__getDetail(postingId));
   }, [dispatch, postingId, isSuccess]);
 
-  // if (isLoading) return <Spinners />;
-  // if (error) return <div>{error.msg}</div>;
+  if (isLoading) return <Spinners />;
+
+  if (error) return <div>{error.msg}</div>;
 
   return (
     <StDetailWrap>
@@ -76,48 +48,16 @@ const Detail = () => {
         <ElImgWrap>
           <img src={details.image} alt="" />
         </ElImgWrap>
-        <StCreatorProfile>
-          <div>
-            {/* 프로필 이미지 */}
-            <div className="profileWrap">
-              {details.profileImage !== "" ? (
-                <img
-                  src={details.profileImage}
-                  alt=""
-                  onClick={() => onClickMoveProfileHandler(details.memberId)}
-                />
-              ) : (
-                <img
-                  src={profileDefault}
-                  alt=""
-                  onClick={() => onClickMoveProfileHandler(details.memberId)}
-                />
-              )}
-            </div>
-            {/* 유저 정보 */}
-            <DetailSpan
-              titleText={details.nickname}
-              bodyText={details.address?.split(" ", 3).join(" ")}
-              margin="0 0 0.25rem"
-              fontSize="0.875rem"
-            />
-          </div>
-          <div className="postingOption">
-            {postingModal && <DetailPostingOptionModal postingId={postingId} />}
-            {details.myPosting && (
-              <DetailMoreButton onClick={onClickPostingModalHandler} />
-            )}
-          </div>
-        </StCreatorProfile>
+        <DetailCreatorProfile
+          details={details}
+          postingId={postingId}
+          onClickMoveProfileHandler={onClickMoveProfileHandler}
+          DetailSpan={DetailSpan}
+          DetailMoreButton={DetailMoreButton}
+        />
         <hr />
-        <StContent>
-          <h3>{details.title}</h3>
-          {/* <h3>{longtitleForm(details.title)}</h3> */}
-          <span>
-            <u>{details.category}</u> {details.createdAt?.split(" ")[0]}
-          </span>
-          <p>{details.content}</p>
-        </StContent>
+        <DetailContent details={details} />
+
         <hr />
         <DetailApplicationBody details={details} />
         <hr />
@@ -132,74 +72,19 @@ const Detail = () => {
         />
         <StBuyLocation id="map" />
         <hr />
-        <ElApplicationWrap>
-          {details.doneStatus ? (
-            <ButtonBasic background="#adadad" color="white">
-              <img src={applicateBtnIcon} alt="" /> 완료된 게시물입니다.
-            </ButtonBasic>
-          ) : (
-            <>
-              {details.myPosting ? (
-                <>
-                  {applicationModal && (
-                    <DetailApplicationList
-                      postingId={postingId}
-                      onClickMoveProfileHandler={onClickMoveProfileHandler}
-                    />
-                  )}
-                  <ButtonBasic _onClick={onClickApplicationModalHandler}>
-                    <img src={applicateBtnIcon} alt="" /> 신청 리스트 보기
-                  </ButtonBasic>
-                </>
-              ) : (
-                <ButtonBasic _onClick={onClickApplicateHandler}>
-                  <img src={applicateBtnIcon} alt="" /> 참가 신청 하기
-                </ButtonBasic>
-                // 참가 확정인 건 어떻게 판단?
-                // <ButtonBasic _onClick={onClickApplicateHandler}>
-                //   <img src={applicateBtnIcon} alt="" /> 참가 취소 하기
-                // </ButtonBasic>
-                // 얼럿 들어가고
-                // postingId하고 토큰
-              )}
-            </>
-          )}
-        </ElApplicationWrap>
-        <StCommentWrap>
-          {commentsLength !== 0 ? (
-            <span> 댓글 {commentsLength}</span>
-          ) : (
-            <span> 댓글 0</span>
-          )}
-          {tokenValue && !details.doneStatus && <DetailCommentForm />}
-        </StCommentWrap>
-        {commentsLength !== 0 &&
-          details.commentList?.map((comment, idx) => (
-            <StCommentList key={comment.nickname[0] + idx}>
-              <div>
-                <span>
-                  <span>
-                    {comment.nickname}
-                    &nbsp;&nbsp;&nbsp;
-                  </span>
-                  <span>{comment.createdAt?.split(" ", 1)}</span>
-                  <p>{comment.content}</p>
-                </span>
-                {tokenValue && (
-                  <div className="commentOption">
-                    {commentModal && (
-                      <DetailCommentModal id={comment.commentId} />
-                    )}
-                    <DetailMoreButton
-                      onClick={onClickCommentModalHandler}
-                      id={comment.commentId}
-                    />
-                  </div>
-                )}
-              </div>
-              <hr key={"hr" + idx} />
-            </StCommentList>
-          ))}
+        {/* 참가 버튼 */}
+        <DetailApplicationBtns
+          details={details}
+          postingId={postingId}
+          tokenValue={tokenValue}
+          onClickMoveProfileHandler={onClickMoveProfileHandler}
+        />
+
+        <DetailCommentList
+          details={details}
+          tokenValue={tokenValue}
+          DetailMoreButton={DetailMoreButton}
+        />
       </StDetailForm>
     </StDetailWrap>
   );
@@ -240,6 +125,7 @@ const ElImgWrap = styled.div`
     width: 100%;
     height: 31.625rem;
     object-fit: cover;
+    /* border: 0.1rem solid ${({ theme }) => theme.colors.grayWeak}; */
     border-radius: 0.5rem;
 
     @media screen and (max-width: 48rem) {
@@ -252,62 +138,6 @@ const ElImgWrap = styled.div`
   }
 `;
 
-const StCreatorProfile = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  /* height: 5rem; */
-  div:first-of-type {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-  }
-  span:nth-of-type(2) {
-    color: ${({ theme }) => theme.colors.grayStrong};
-    font-size: ${({ theme }) => theme.fontSize.sm};
-  }
-  /* 프로필 이미지 래퍼 */
-  .profileWrap {
-    width: 3.725rem;
-    height: 3.725rem;
-    ${({ theme }) => theme.common.flexCenter}
-    margin-right: 0.625rem;
-    border-radius: 5rem;
-    overflow: hidden;
-    border: 0.1rem solid ${({ theme }) => theme.colors.grayWeak};
-  }
-
-  /* 프로필 이미지 */
-  .profileWrap img {
-    flex-shrink: 0;
-    width: 100%;
-    min-height: 100%;
-  }
-
-  .postingOption {
-    position: relative;
-  }
-`;
-
-const StContent = styled.div`
-  h3 {
-    font-weight: 600;
-    font-size: ${({ theme }) => theme.fontSize.xl};
-    margin-bottom: ${({ theme }) => theme.lineHeight.perParagraph};
-  }
-  span {
-    display: inline-block;
-    font-size: ${({ theme }) => theme.fontSize.sm};
-    margin-bottom: ${({ theme }) => theme.lineHeight.perParagraph};
-    color: ${({ theme }) => theme.colors.grayMid};
-  }
-
-  p {
-    line-height: ${({ theme }) => theme.lineHeight.perParagraph};
-  }
-`;
-
 const StBuyLocation = styled.div`
   width: 100%;
   height: 18.725rem;
@@ -317,49 +147,5 @@ const StBuyLocation = styled.div`
 
   @media screen and (max-width: 23.5rem) {
     height: 9.5rem;
-  }
-`;
-
-const ElApplicationWrap = styled.div`
-  position: relative;
-  button {
-    height: 3.125rem;
-    margin: 1.875rem 0;
-  }
-  img {
-    margin-right: 0.4rem;
-  }
-`;
-
-const StCommentWrap = styled.div`
-  margin-bottom: 2rem;
-  div:first-of-type {
-    display: flex;
-    flex-direction: column;
-    background: ${({ theme }) => theme.colors.grayList};
-    padding: 1rem;
-  }
-`;
-
-const StCommentList = styled.div`
-  div {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    font-size: ${({ theme }) => theme.fontSize.sm};
-  }
-  span > span:first-of-type {
-    display: inline-block;
-    margin-bottom: ${({ theme }) => theme.lineHeight.perSpan};
-    font-weight: 600;
-  }
-  span > span:nth-of-type(2) {
-    color: ${({ theme }) => theme.colors.grayMid};
-    font-size: ${({ theme }) => theme.fontSize.xs};
-  }
-
-  .commentOption {
-    position: relative;
   }
 `;
