+++ src/redux/modules/application/applicationSlice.js
@@ -1,13 +1,15 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { baseURL, baseURLwToken } from "../../../core/axios";
+import { baseURLwToken } from "../../../core/axios";
 
 const initialState = {
   postApplication: {},
   getApplication: [],
   allowApplication: {},
   denyApplication: {},
+  cancelApplication: {},
   isLoading: false,
   error: null,
+  isSuccess: false,
 };
 
 export const __getApplication = createAsyncThunk(
@@ -31,8 +33,7 @@ export const __postApplication = createAsyncThunk(
       const { data } = await baseURLwToken.post(
         `participant/posting/${postingId}`
       );
-      alert(data.msg);
-      // return thunkAPI.fulfillWithValue(data);
+      return thunkAPI.fulfillWithValue(data.msg);
     } catch (err) {
       return thunkAPI.rejectWithValue(err);
     }
@@ -46,8 +47,7 @@ export const __allowApplication = createAsyncThunk(
       const { data } = await baseURLwToken.post(
         `participant/${applicationId}/posting/${postingId}`
       );
-      alert(data.msg);
-      // return thunkAPI.fulfillWithValue(data)
+      return thunkAPI.fulfillWithValue(data.msg);
     } catch (err) {
       return thunkAPI.rejectWithValue(err);
     }
@@ -61,8 +61,21 @@ export const __denyApplication = createAsyncThunk(
       const { data } = await baseURLwToken.delete(
         `participant/${applicationId}/posting/${postingId}`
       );
-      alert(data.msg);
-      // return thunkAPI.fulfillWithValue(data)
+      return thunkAPI.fulfillWithValue(data.msg);
+    } catch (err) {
+      return thunkAPI.rejectWithValue(err);
+    }
+  }
+);
+
+export const __cancelApplication = createAsyncThunk(
+  "application/cancel",
+  async (payload, thunkAPI) => {
+    try {
+      const { data } = await baseURLwToken.delete(
+        `participant/posting/${payload}`
+      );
+      return thunkAPI.fulfillWithValue(data.msg);
     } catch (err) {
       return thunkAPI.rejectWithValue(err);
     }
@@ -88,6 +101,10 @@ export const applicationSlice = createSlice({
         state.allowApplication = action.payload;
       })
       .addCase(__denyApplication.fulfilled, (state, action) => {
+        state.isLoading = false;
+        state.denyApplication = action.payload;
+      })
+      .addCase(__cancelApplication.fulfilled, (state, action) => {
         state.isLoading = false;
         state.denyApplication = action.payload;
       });
